Add ConcurrencyLimiter middleware to control parallel requests
We introduced a ConcurrencyLimiter middleware to prevent too many parallel
requests from being processed simultaneously. Our requests are large in
volume and all data is written to a single sink at the end of processing,
which leads to high memory usage if too many requests are in-flight.

By limiting concurrency, we avoid long queue buildup that could cause
excessive memory consumption and potentially trigger OOMKill on the server.